move update_exchange_rates to accounting/tasks.py
