+++ corehq/apps/accounting/tasks.py
@@ -1,4 +1,6 @@
 import datetime
+import json
+import urllib2
 from StringIO import StringIO
 from urllib import urlencode
 
@@ -25,6 +27,7 @@ from corehq.apps.accounting.exceptions import (
 )
 from corehq.apps.accounting.invoicing import DomainInvoiceFactory
 from corehq.apps.accounting.models import (
+    Currency,
     Subscription, Invoice,
     SubscriptionAdjustment, SubscriptionAdjustmentReason,
     SubscriptionAdjustmentMethod,
@@ -502,3 +505,22 @@ def weekly_digest():
 def pay_autopay_invoices():
     """ Check for autopayable invoices every day and pay them """
     AutoPayInvoicePaymentHandler().pay_autopayable_invoices(datetime.datetime.today())
+
+
+@periodic_task(run_every=crontab(minute=0, hour=0), queue='background_queue')
+def update_exchange_rates(app_id=settings.OPEN_EXCHANGE_RATES_API_ID):
+    try:
+        log_accounting_info("Updating exchange rates...")
+        rates = json.load(urllib2.urlopen(
+            'https://openexchangerates.org/api/latest.json?app_id=%s' % app_id))['rates']
+        default_rate = float(rates[Currency.get_default().code])
+        for code, rate in rates.items():
+            currency, _ = Currency.objects.get_or_create(code=code)
+            currency.rate_to_default = float(rate) / default_rate
+            currency.save()
+            log_accounting_info("Exchange rate for %(code)s updated %(rate)f." % {
+                'code': currency.code,
+                'rate': currency.rate_to_default,
+            })
+    except Exception as e:
+        log_accounting_error(e.message)
